+++ TFT/src/User/Menu/ABL.h
@@ -10,10 +10,11 @@ extern "C" {
 // called by parseAck() to notify ABL process status
 void ablUpdateStatus(bool succeeded);
 
+// Start ABL process
+void ablStart(void);
+
 void menuUBLSave(void);
 void menuUBLLoad(void);
-
-void menuABL(void);
 
 #ifdef __cplusplus
 }
