+++ TFT/src/User/Menu/BedLeveling.c
@@ -29,13 +29,12 @@ void menuBedLeveling(void)
       {ICON_LEVELING_OFF,            LABEL_BL_DISABLE},
       {ICON_Z_FADE,                  LABEL_ABL_Z},
       {ICON_BACKGROUND,              LABEL_BACKGROUND},
-      {ICON_BACKGROUND,              LABEL_BACKGROUND},
+      {ICON_HEAT_FAN,                LABEL_UNIFIEDHEAT},
       {ICON_BACK,                    LABEL_BACK},
     }
   };
 
   KEY_VALUES key_num = KEY_IDLE;
-  void (*menuBL)(void) = menuABL;
   int8_t levelStateOld = -1;
 
   switch (infoMachineSettings.leveling)
@@ -53,8 +52,6 @@ void menuBedLeveling(void)
     case BL_MBL:
       bedLevelingItems.title.index = LABEL_MBL_SETTINGS;
       bedLevelingItems.items[0].label.index = LABEL_MBL;
-
-      menuBL = menuMBL;
       break;
 
     default:
@@ -71,12 +68,6 @@ void menuBedLeveling(void)
   {
     bedLevelingItems.items[5].icon = ICON_PROBE_OFFSET;
     bedLevelingItems.items[5].label.index = LABEL_P_OFFSET;
-  
-    if (infoSettings.z_steppers_alignment)
-    {
-      bedLevelingItems.items[6].icon = ICON_Z_ALIGN;
-      bedLevelingItems.items[6].label.index = LABEL_Z_ALIGN;
-    }
   }
 
   menuDrawPage(&bedLevelingItems);
@@ -87,7 +78,7 @@ void menuBedLeveling(void)
     switch (key_num)
     {
       case KEY_ICON_0:
-        infoMenu.menu[++infoMenu.cur] = menuBL;
+        infoMenu.menu[++infoMenu.cur] = menuBedLevelingLayer2;
         break;
 
       case KEY_ICON_1:
@@ -123,11 +114,19 @@ void menuBedLeveling(void)
         break;
 
       case KEY_ICON_6:
-        if (infoMachineSettings.zProbe == ENABLED && infoSettings.z_steppers_alignment)
-          storeCmd("G34\n");
+        infoMenu.menu[++infoMenu.cur] = menuUnifiedHeat;
         break;
 
       case KEY_ICON_7:
+        for (uint8_t i = 0; i < MAX_HEATER_COUNT; i++)
+        {
+          if (heatGetTargetTemp(i) > 0)
+          {
+            setDialogText(LABEL_WARNING, LABEL_HEATERS_ON, LABEL_CONFIRM, LABEL_CANCEL);
+            showDialog(DIALOG_TYPE_QUESTION, heatCoolDown, NULL, NULL);
+            break;
+          }
+        }
         infoMenu.cur--;
         break;
