--- TFT/src/User/Menu/ManualLeveling.h
@@ -1,14 +1,0 @@
-#ifndef _MANUAL_LEVELING_H_
-#define _MANUAL_LEVELING_H_
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void menuManualLeveling(void);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
