+++ TFT/src/User/Menu/Tuning.c
@@ -8,7 +8,7 @@ const MENUITEMS TuningItems = {
   {
     {ICON_PID,                     LABEL_PID},
     {ICON_TUNE_EXTRUDER,           LABEL_TUNE_EXTRUDER},
-    {ICON_BACKGROUND,              LABEL_BACKGROUND},
+    {ICON_PROBE_OFFSET,            LABEL_H_OFFSET},
     {ICON_BACKGROUND,              LABEL_BACKGROUND},
     {ICON_BACKGROUND,              LABEL_BACKGROUND},
     {ICON_BACKGROUND,              LABEL_BACKGROUND},
@@ -34,6 +34,12 @@ void menuTuning(void)
 
       case KEY_ICON_1:
         infoMenu.menu[++infoMenu.cur] = menuTuneExtruder;
+        break;
+
+      case KEY_ICON_2:
+        storeCmd("M206\n");
+        zOffsetSetMenu(false);  // use Home Offset menu
+        infoMenu.menu[++infoMenu.cur] = menuZOffset;
         break;
 
       case KEY_ICON_7:
