+++ TFT/src/User/Menu/UnifiedMove.c
@@ -21,6 +21,12 @@ void menuUnifiedMove(void)
 
   KEY_VALUES key_num = KEY_IDLE;
 
+  if (infoMachineSettings.leveling != BL_DISABLED)
+  {
+    UnifiedMoveItems.items[6].icon = ICON_LEVELING;
+    UnifiedMoveItems.items[6].label.index = LABEL_BED_LEVELING;
+  }
+
   menuDrawPage(&UnifiedMoveItems);
 
   while (infoMenu.menu[infoMenu.cur] == menuUnifiedMove)
@@ -49,7 +55,12 @@ void menuUnifiedMove(void)
         break;
 
       case KEY_ICON_5:
-        infoMenu.menu[++infoMenu.cur] = menuLeveling;
+        infoMenu.menu[++infoMenu.cur] = menuManualLeveling;
+        break;
+
+      case KEY_ICON_6:
+        if (infoMachineSettings.leveling != BL_DISABLED)
+          infoMenu.menu[++infoMenu.cur] = menuBedLeveling;
         break;
 
       case KEY_ICON_7:
