+++ TFT/src/User/includes.h
@@ -107,8 +107,8 @@
 #include "FeatureSettings.h"
 #include "SendGcode.h"
 #include "Leveling.h"
-#include "ManualLeveling.h"
 #include "BedLeveling.h"
+#include "BedLevelingLayer2.h"
 #include "MBL.h"
 #include "ABL.h"
 #include "BLTouch.h"
